+++ Modules/Intent.Modules.EntityFrameworkCore/release-notes.md
@@ -1,5 +1,6 @@
 ﻿### Version 5.0.6
 
+- Improved: Set your default schema name for your `ApplicationDbContext` and your HistoryTable.
 - Fixed: Checking for types with serialization will also include Record types.
 
 ### Version 5.0.5